requirements.txt
Flask==0.10.1
Flask-RESTful==0.2.12
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
aniso8601==0.82
itsdangerous==0.24
pyserial==2.7
pytz==2014.4
six==1.7.2
